compilation warning fixed
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@6184 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/LocalPlayer.cxx
@@ -237,7 +237,7 @@ void			LocalPlayer::doUpdateMotion(float dt)
 
       // reset our flap count if we have wings
       if (getFlag() == Flags::Wings)
-         wingsFlapCount = BZDB.eval(StateDatabase::BZDB_WINGSJUMPCOUNT);
+         wingsFlapCount = (int) BZDB.eval(StateDatabase::BZDB_WINGSJUMPCOUNT);
 
       if ((oldPosition[2] < 0.0f) && (getFlag() == Flags::Burrow))
 	newVelocity[2] += 4 * BZDB.eval(StateDatabase::BZDB_GRAVITY) * dt;
